feat: add armor constants, loadout hook and components
